cabana: split utils/util into Qt-free util and qtutil (#38723)
