+++ openpilot/tools/cabana/utils/util.cc
@@ -1,18 +1,19 @@
 #include "tools/cabana/utils/util.h"
 
 #include <algorithm>
+#include <cctype>
 #include <cerrno>
-#include <cmath>
 #include <cstdint>
 #include <cstdio>
 #include <cstdlib>
 #include <csignal>
+#include <cstring>
 #include <filesystem>
 #include <limits>
-#include <memory>
 #include <mutex>
 #include <string>
 #include <thread>
+#include <unordered_map>
 #include <sys/socket.h>
 #include <sys/wait.h>
 #include <unistd.h>
@@ -20,11 +21,6 @@
 #include <mach-o/dyld.h>
 #endif
 
-#include <QColor>
-#include <QFontDatabase>
-#include <QPixmapCache>
-#include <QPainterPath>
-#include <unordered_map>
 #include "common/util.h"
 
 static const std::thread::id main_thread_id = std::this_thread::get_id();
@@ -53,22 +49,21 @@ void utils::drainMainThreadQueue() {
 
 // SegmentTree
 
-void SegmentTree::build(const std::vector<QPointF> &arr) {
-  size = arr.size();
+void SegmentTree::build(int n, const std::function<double(int)> &y) {
+  size = n;
   tree.resize(4 * size);  // size of the tree is 4 times the size of the array
   if (size > 0) {
-    build_tree(arr, 1, 0, size - 1);
+    build_tree(y, 1, 0, size - 1);
   }
 }
 
-void SegmentTree::build_tree(const std::vector<QPointF> &arr, int n, int left, int right) {
+void SegmentTree::build_tree(const std::function<double(int)> &y, int n, int left, int right) {
   if (left == right) {
-    const double y = arr[left].y();
-    tree[n] = {y, y};
+    tree[n] = {y(left), y(left)};
   } else {
     const int mid = (left + right) >> 1;
-    build_tree(arr, 2 * n, left, mid);
-    build_tree(arr, 2 * n + 1, mid + 1, right);
+    build_tree(y, 2 * n, left, mid);
+    build_tree(y, 2 * n + 1, mid + 1, right);
     tree[n] = {std::min(tree[2 * n].first, tree[2 * n + 1].first), std::max(tree[2 * n].second, tree[2 * n + 1].second)};
   }
 }
@@ -84,119 +79,22 @@ std::pair<double, double> SegmentTree::get_minmax(int n, int left, int right, in
   return {std::min(l.first, r.first), std::max(l.second, r.second)};
 }
 
-// MessageBytesDelegate
-
-MessageBytesDelegate::MessageBytesDelegate(QObject *parent, bool multiple_lines)
-    : font_metrics(QApplication::font()), multiple_lines(multiple_lines), QStyledItemDelegate(parent) {
-  fixed_font = QFontDatabase::systemFont(QFontDatabase::FixedFont);
-  byte_size = QFontMetrics(fixed_font).size(Qt::TextSingleLine, "00 ") + QSize(0, 2);
-  for (int i = 0; i < 256; ++i) {
-    hex_text_table[i].setText(QStringLiteral("%1").arg(i, 2, 16, QLatin1Char('0')).toUpper());
-    hex_text_table[i].prepare({}, fixed_font);
-  }
-  h_margin = QApplication::style()->pixelMetric(QStyle::PM_FocusFrameHMargin) + 1;
-  v_margin = QApplication::style()->pixelMetric(QStyle::PM_FocusFrameVMargin) + 1;
-}
-
-QSize MessageBytesDelegate::sizeForBytes(int n) const {
-  int rows = multiple_lines ? std::max(1, n / 8) : 1;
-  return {(n / rows) * byte_size.width() + h_margin * 2, rows * byte_size.height() + v_margin * 2};
-}
-
-QSize MessageBytesDelegate::sizeHint(const QStyleOptionViewItem &option, const QModelIndex &index) const {
-  auto data = index.data(BytesRole);
-  return sizeForBytes(data.isValid() ? static_cast<std::vector<uint8_t> *>(data.value<void *>())->size() : 0);
-}
-
-void MessageBytesDelegate::paint(QPainter *painter, const QStyleOptionViewItem &option, const QModelIndex &index) const {
-  if (option.state & QStyle::State_Selected) {
-    painter->fillRect(option.rect, option.palette.brush(QPalette::Normal, QPalette::Highlight));
-  }
-
-  QRect item_rect = option.rect.adjusted(h_margin, v_margin, -h_margin, -v_margin);
-  QColor highlighted_color = option.palette.color(QPalette::HighlightedText);
-  auto text_color = index.data(Qt::ForegroundRole).value<QColor>();
-  bool inactive = text_color.isValid();
-  if (!inactive) {
-    text_color = option.palette.color(QPalette::Text);
-  }
-  auto data = index.data(BytesRole);
-  if (!data.isValid()) {
-    painter->setFont(option.font);
-    painter->setPen(option.state & QStyle::State_Selected ? highlighted_color : text_color);
-    QString text = font_metrics.elidedText(index.data(Qt::DisplayRole).toString(), Qt::ElideRight, item_rect.width());
-    painter->drawText(item_rect, Qt::AlignLeft | Qt::AlignVCenter, text);
-    return;
-  }
-
-  // Paint hex column
-  const auto &bytes = *static_cast<std::vector<uint8_t> *>(data.value<void *>());
-  const auto &colors = *static_cast<std::vector<CabanaColor> *>(index.data(ColorsRole).value<void *>());
-
-  painter->setFont(fixed_font);
-  const QPen text_pen(option.state & QStyle::State_Selected ? highlighted_color : text_color);
-  const QPoint pt = item_rect.topLeft();
-  for (int i = 0; i < bytes.size(); ++i) {
-    int row = !multiple_lines ? 0 : i / 8;
-    int column = !multiple_lines ? i : i % 8;
-    QRect r({pt.x() + column * byte_size.width(), pt.y() + row * byte_size.height()}, byte_size);
-
-    if (!inactive && i < colors.size() && colors[i].alpha() > 0) {
-      if (option.state & QStyle::State_Selected) {
-        painter->setPen(option.palette.color(QPalette::Text));
-        painter->fillRect(r, option.palette.color(QPalette::Window));
-      }
-      painter->fillRect(r, toQColor(colors[i]));
-    } else {
-      painter->setPen(text_pen);
-    }
-    utils::drawStaticText(painter, r, hex_text_table[bytes[i]]);
-  }
-}
-
-// TabBar
-
-int TabBar::addTab(const QString &text) {
-  int index = QTabBar::addTab(text);
-  QToolButton *btn = new ToolButton("x", tr("Close Tab"));
-  int width = style()->pixelMetric(QStyle::PM_TabCloseIndicatorWidth, nullptr, btn);
-  int height = style()->pixelMetric(QStyle::PM_TabCloseIndicatorHeight, nullptr, btn);
-  btn->setFixedSize({width, height});
-  setTabButton(index, QTabBar::RightSide, btn);
-  QObject::connect(btn, &QToolButton::clicked, this, &TabBar::closeTabClicked);
-  return index;
-}
-
-void TabBar::closeTabClicked() {
-  QObject *object = sender();
-  for (int i = 0; i < count(); ++i) {
-    if (tabButton(i, QTabBar::RightSide) == object) {
-      emit tabCloseRequested(i);
-      break;
-    }
-  }
-}
-
 // UnixSignalHandler
 
-UnixSignalHandler::UnixSignalHandler() {
+UnixSignalHandler::UnixSignalHandler(std::function<void()> on_signal) {
   if (::socketpair(AF_UNIX, SOCK_STREAM, 0, sig_fd)) {
-    qFatal("Couldn't create TERM socketpair");
+    fprintf(stderr, "Couldn't create TERM socketpair\n");
+    abort();
   }
 
-  waiter = std::thread([this]() {
+  waiter = std::thread([this, on_signal = std::move(on_signal)]() {
     int tmp = 0;
     while (::read(sig_fd[1], &tmp, sizeof(tmp)) < 0) {
       if (errno != EINTR) return;
     }
     if (shutting_down.load()) return;
 
-    // Marshal exit onto the GUI thread (qApp methods are not thread-safe).
-    QMetaObject::invokeMethod(qApp, []() {
-      printf("\nexiting...\n");
-      qApp->closeAllWindows();
-      qApp->exit();
-    }, Qt::QueuedConnection);
+    on_signal();
   });
 
   std::signal(SIGINT, signalHandler);
@@ -216,118 +114,131 @@ void UnixSignalHandler::signalHandler(int s) {
   (void)!::write(sig_fd[0], &s, sizeof(s));
 }
 
-// NameValidator
+// validators
 
-NameValidator::NameValidator(QObject *parent) : QValidator(parent) {}
-
-QValidator::State NameValidator::validate(QString &input, int &pos) const {
-  Q_UNUSED(pos);
-  input.replace(' ', '_');
-  if (input.isEmpty()) return QValidator::Intermediate;
-  for (const QChar &c : input) {
-    if (!c.isLetterOrNumber() && c != '_') return QValidator::Invalid;
+ValidState validateName(std::string &input) {
+  std::replace(input.begin(), input.end(), ' ', '_');
+  if (input.empty()) return ValidState::Intermediate;
+  for (const unsigned char c : input) {
+    if (!std::isalnum(c) && c != '_') return ValidState::Invalid;
   }
-  return QValidator::Acceptable;
+  return ValidState::Acceptable;
 }
 
-// NodeValidator
-
-NodeValidator::NodeValidator(QObject *parent) : QValidator(parent) {}
-
-QValidator::State NodeValidator::validate(QString &input, int &pos) const {
-  Q_UNUSED(pos);
-  if (input.isEmpty()) return QValidator::Intermediate;
+ValidState validateNodes(const std::string &input) {
+  if (input.empty()) return ValidState::Intermediate;
   // Match ^\w+(,\w+)*$ ; a trailing comma is Intermediate (user still typing).
   bool need_word = true;
-  for (const QChar &c : input) {
-    if (c.isLetterOrNumber() || c == '_') {
+  for (const unsigned char c : input) {
+    if (std::isalnum(c) || c == '_') {
       need_word = false;
     } else if (c == ',' && !need_word) {
       need_word = true;
     } else {
-      return QValidator::Invalid;
+      return ValidState::Invalid;
     }
   }
-  return need_word ? QValidator::Intermediate : QValidator::Acceptable;
+  return need_word ? ValidState::Intermediate : ValidState::Acceptable;
 }
 
-// NonWhitespaceValidator
-
-NonWhitespaceValidator::NonWhitespaceValidator(QObject *parent) : QValidator(parent) {}
-
-QValidator::State NonWhitespaceValidator::validate(QString &input, int &pos) const {
-  Q_UNUSED(pos);
-  if (input.isEmpty()) return QValidator::Intermediate;
-  for (const QChar &c : input) {
-    if (c.isSpace()) return QValidator::Invalid;
+ValidState validateNonWhitespace(const std::string &input) {
+  if (input.empty()) return ValidState::Intermediate;
+  for (const unsigned char c : input) {
+    if (std::isspace(c)) return ValidState::Invalid;
   }
-  return QValidator::Acceptable;
+  return ValidState::Acceptable;
 }
 
-// IpAddressValidator
-
-IpAddressValidator::IpAddressValidator(QObject *parent) : QValidator(parent) {}
-
-QValidator::State IpAddressValidator::validate(QString &input, int &pos) const {
-  Q_UNUSED(pos);
-  if (input.isEmpty()) return QValidator::Intermediate;
+ValidState validateIpAddress(const std::string &input) {
+  if (input.empty()) return ValidState::Intermediate;
 
   int dots = 0;
   int value = 0;
   bool has_digit = false;
-  for (const QChar &c : input) {
-    if (c.isDigit()) {
-      value = has_digit ? value * 10 + c.digitValue() : c.digitValue();
-      if (value > 255) return QValidator::Invalid;
+  for (const unsigned char c : input) {
+    if (std::isdigit(c)) {
+      value = has_digit ? value * 10 + (c - '0') : (c - '0');
+      if (value > 255) return ValidState::Invalid;
       has_digit = true;
     } else if (c == '.') {
-      if (!has_digit || dots >= 3) return QValidator::Invalid;
+      if (!has_digit || dots >= 3) return ValidState::Invalid;
       ++dots;
       has_digit = false;
       value = 0;
     } else {
-      return QValidator::Invalid;
+      return ValidState::Invalid;
     }
   }
-  return (dots == 3 && has_digit) ? QValidator::Acceptable : QValidator::Intermediate;
+  return (dots == 3 && has_digit) ? ValidState::Acceptable : ValidState::Intermediate;
 }
 
-DoubleValidator::DoubleValidator(QObject *parent) : QValidator(parent) {}
+ValidState validateDouble(const std::string &input) {
+  if (input.empty()) return ValidState::Intermediate;
 
-QValidator::State DoubleValidator::validate(QString &input, int &pos) const {
-  Q_UNUSED(pos);
-  if (input.isEmpty()) return QValidator::Intermediate;
-
-  // Match QString::toDouble(): C locale, no hex floats / inf / nan.
-  const std::string bytes = input.toLatin1().toStdString();
-  // strtod accepts 0x… hex floats and p-exponents; QString::toDouble does not.
-  if (bytes.find_first_of("xXpP") != std::string::npos) {
-    return QValidator::Invalid;
+  // C locale, no hex floats / p-exponents / inf / nan (strtod accepts them, the DBC parser does not)
+  if (input.find_first_of("xXpP") != std::string::npos) {
+    return ValidState::Invalid;
   }
 
-  const char *start = bytes.c_str();
+  const char *start = input.c_str();
   char *end = nullptr;
   const double value = std::strtod(start, &end);
   if (end == start) {
     // Still typing a sign, decimal point, or exponent prefix.
     if (input == "-" || input == "+" || input == "." || input == "-." || input == "+.") {
-      return QValidator::Intermediate;
+      return ValidState::Intermediate;
     }
-    return QValidator::Invalid;
+    return ValidState::Invalid;
   }
   if (*end == '\0') {
-    // Reject inf/nan (strtod accepts them; QDoubleValidator / toDouble path should not).
-    return std::isfinite(value) ? QValidator::Acceptable : QValidator::Invalid;
+    return std::isfinite(value) ? ValidState::Acceptable : ValidState::Invalid;
   }
 
   // Partial exponent / trailing sign while typing (e.g. "1e", "1e-", "1.").
   for (const char *p = end; *p; ++p) {
     const char c = *p;
     if (!(c == 'e' || c == 'E' || c == '+' || c == '-' || c == '.' || (c >= '0' && c <= '9'))) {
-      return QValidator::Invalid;
+      return ValidState::Invalid;
     }
   }
-  return QValidator::Intermediate;
+  return ValidState::Intermediate;
+}
+
+// embedded at build time from the bootstrap_icons package (see SConscript)
+extern const unsigned char bootstrap_icons_svg[];
+extern const size_t bootstrap_icons_svg_len;
+
+static std::unordered_map<std::string, std::string> load_bootstrap_icons() {
+  std::unordered_map<std::string, std::string> icons;
+
+  const std::string content(reinterpret_cast<const char *>(bootstrap_icons_svg), bootstrap_icons_svg_len);
+  const std::string sym_open = "<symbol ";
+  const std::string sym_close = "</symbol>";
+  const std::string id_attr = "id=\"";
+
+  size_t pos = 0;
+  while ((pos = content.find(sym_open, pos)) != std::string::npos) {
+    size_t end = content.find(sym_close, pos);
+    if (end == std::string::npos) break;
+    end += sym_close.size();
+
+    // extract id
+    size_t id_start = content.find(id_attr, pos);
+    if (id_start != std::string::npos && id_start < end) {
+      id_start += id_attr.size();
+      size_t id_end = content.find('"', id_start);
+      if (id_end != std::string::npos && id_end < end) {
+        std::string id = content.substr(id_start, id_end - id_start);
+        std::string svg_str = content.substr(pos, end - pos);
+        // replace <symbol with <svg, </symbol> with </svg>
+        svg_str.replace(0, 7, "<svg");               // "<symbol" (7) -> "<svg" (4)
+        svg_str.replace(svg_str.size() - 9, 9, "</svg>");  // "</symbol>" (9) -> "</svg>" (6)
+        icons[id] = std::move(svg_str);
+      }
+    }
+    pos = end;
+  }
+  return icons;
 }
 
 namespace utils {
@@ -384,77 +295,19 @@ bool setClipboardText(const std::string &text) {
   return false;
 }
 
-bool isDarkTheme() {
-  QColor windowColor = QApplication::palette().color(QPalette::Window);
-  return windowColor.lightness() < 128;
-}
-
-QPixmap icon(const QString &id) {
-  bool dark_theme = isDarkTheme();
-
-  QPixmap pm;
-  QString key = "bootstrap_" % id % (dark_theme ? "1" : "0");
-  if (!QPixmapCache::find(key, &pm)) {
-    pm = bootstrapPixmap(id);
-    if (dark_theme) {
-      QPainter p(&pm);
-      p.setCompositionMode(QPainter::CompositionMode_SourceIn);
-      p.fillRect(pm.rect(), QColor("#bbbbbb"));
-    }
-    QPixmapCache::insert(key, pm);
-  }
-  return pm;
-}
-
-void setTheme(int theme) {
-  auto style = QApplication::style();
-  if (!style) return;
-
-  static int prev_theme = 0;
-  if (theme != prev_theme) {
-    prev_theme = theme;
-    QPalette new_palette;
-    if (theme == DARK_THEME) {
-      // "Darcula" like dark theme
-      new_palette.setColor(QPalette::Window, QColor("#353535"));
-      new_palette.setColor(QPalette::WindowText, QColor("#bbbbbb"));
-      new_palette.setColor(QPalette::Base, QColor("#3c3f41"));
-      new_palette.setColor(QPalette::AlternateBase, QColor("#3c3f41"));
-      new_palette.setColor(QPalette::ToolTipBase, QColor("#3c3f41"));
-      new_palette.setColor(QPalette::ToolTipText, QColor("#bbb"));
-      new_palette.setColor(QPalette::Text, QColor("#bbbbbb"));
-      new_palette.setColor(QPalette::Button, QColor("#3c3f41"));
-      new_palette.setColor(QPalette::ButtonText, QColor("#bbbbbb"));
-      new_palette.setColor(QPalette::Highlight, QColor("#2f65ca"));
-      new_palette.setColor(QPalette::HighlightedText, QColor("#bbbbbb"));
-      new_palette.setColor(QPalette::BrightText, QColor("#f0f0f0"));
-      new_palette.setColor(QPalette::Disabled, QPalette::ButtonText, QColor("#777777"));
-      new_palette.setColor(QPalette::Disabled, QPalette::WindowText, QColor("#777777"));
-      new_palette.setColor(QPalette::Disabled, QPalette::Text, QColor("#777777"));
-      new_palette.setColor(QPalette::Light, QColor("#777777"));
-      new_palette.setColor(QPalette::Dark, QColor("#353535"));
-    } else {
-      new_palette = style->standardPalette();
-    }
-    qApp->setPalette(new_palette);
-    style->polish(qApp);
-    for (auto w : QApplication::allWidgets()) {
-      w->setPalette(new_palette);
-    }
-  }
+std::string bootstrapSvg(const std::string &id) {
+  static auto icons = load_bootstrap_icons();
+  auto it = icons.find(id);
+  return it != icons.end() ? it->second : std::string();
 }
 
 }  // namespace utils
 
 int num_decimals(double num) {
-  const QString string = QString::number(num);
-  auto dot_pos = string.indexOf('.');
-  return dot_pos == -1 ? 0 : string.size() - dot_pos - 1;
-}
-
-void sigTermHandler(int s) {
-  std::signal(s, SIG_DFL);
-  qApp->quit();
+  char buf[32];
+  snprintf(buf, sizeof(buf), "%g", num);
+  const char *dot = strpbrk(buf, ".,");  // Qt sets LC_ALL from the environment so the decimal mark may be a comma
+  return dot ? (int)strlen(dot + 1) : 0;
 }
 
 std::filesystem::path executableDir() {
@@ -468,71 +321,4 @@ std::filesystem::path executableDir() {
 #else
   return std::filesystem::path(util::readlink("/proc/self/exe")).parent_path();
 #endif
-}
-
-void initApp(int argc, char *argv[], bool disable_hidpi) {
-  // setup signal handlers to exit gracefully
-  std::signal(SIGINT, sigTermHandler);
-  std::signal(SIGTERM, sigTermHandler);
-
-#ifdef __APPLE__
-  // Get the devicePixelRatio, and scale accordingly to maintain 1:1 rendering
-  QApplication tmp(argc, argv);
-  if (disable_hidpi) {
-    qputenv("QT_SCALE_FACTOR", QString::number(1.0 / tmp.devicePixelRatio()).toLocal8Bit());
-  }
-#endif
-
-  qputenv("QT_DBL_CLICK_DIST", "150");
-  // ensure the current dir matches the exectuable's directory
-  std::error_code ec;
-  std::filesystem::current_path(executableDir(), ec);
-}
-
-// embedded at build time from the bootstrap_icons package (see SConscript)
-extern const unsigned char bootstrap_icons_svg[];
-extern const size_t bootstrap_icons_svg_len;
-
-static std::unordered_map<std::string, std::string> load_bootstrap_icons() {
-  std::unordered_map<std::string, std::string> icons;
-
-  const std::string content(reinterpret_cast<const char *>(bootstrap_icons_svg), bootstrap_icons_svg_len);
-  const std::string sym_open = "<symbol ";
-  const std::string sym_close = "</symbol>";
-  const std::string id_attr = "id=\"";
-
-  size_t pos = 0;
-  while ((pos = content.find(sym_open, pos)) != std::string::npos) {
-    size_t end = content.find(sym_close, pos);
-    if (end == std::string::npos) break;
-    end += sym_close.size();
-
-    // extract id
-    size_t id_start = content.find(id_attr, pos);
-    if (id_start != std::string::npos && id_start < end) {
-      id_start += id_attr.size();
-      size_t id_end = content.find('"', id_start);
-      if (id_end != std::string::npos && id_end < end) {
-        std::string id = content.substr(id_start, id_end - id_start);
-        std::string svg_str = content.substr(pos, end - pos);
-        // replace <symbol with <svg, </symbol> with </svg>
-        svg_str.replace(0, 7, "<svg");               // "<symbol" (7) -> "<svg" (4)
-        svg_str.replace(svg_str.size() - 9, 9, "</svg>");  // "</symbol>" (9) -> "</svg>" (6)
-        icons[id] = std::move(svg_str);
-      }
-    }
-    pos = end;
-  }
-  return icons;
-}
-
-QPixmap bootstrapPixmap(const QString &id) {
-  static auto icons = load_bootstrap_icons();
-
-  QPixmap pixmap;
-  auto it = icons.find(id.toStdString());
-  if (it != icons.end()) {
-    pixmap.loadFromData((const uchar *)it->second.data(), it->second.size(), "svg");
-  }
-  return pixmap;
 }
