+++ openpilot/tools/cabana/utils/util.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <array>
+#include <algorithm>
 #include <atomic>
 #include <cmath>
 #include <filesystem>
@@ -10,136 +10,68 @@
 #include <vector>
 #include <utility>
 
-#include <QApplication>
-#include <QColor>
-#include <QComboBox>
-#include <QDialog>
-#include <QFont>
-#include <QFontMetrics>
-#include <QGroupBox>
-#include <QLineEdit>
-#include <QPainter>
-#include <QSpinBox>
-#include <QStaticText>
-#include <QStringBuilder>
-#include <QStyledItemDelegate>
-#include <QToolButton>
-#include <QValidator>
+#include "tools/cabana/core/color.h"
 
-#include "tools/cabana/core/observable.h"
-#include "tools/cabana/dbc/dbc.h"
-#include "tools/cabana/settings.h"
-#include "tools/cabana/utils/strings.h"
-
-// needed by QVariant::fromValue() in the Qt views; goes away with QVariant
-Q_DECLARE_METATYPE(MessageId)
-Q_DECLARE_METATYPE(ValueDescription)
-
-inline QColor toQColor(const CabanaColor &color) {
-  return QColor(color.r, color.g, color.b, color.a);
-}
-
-class LogSlider : public QSlider {
-  Q_OBJECT
-
+class SegmentTree {
 public:
-  LogSlider(double factor, Qt::Orientation orientation, QWidget *parent = nullptr) : factor(factor), QSlider(orientation, parent) {}
+  SegmentTree() = default;
+  void build(int n, const std::function<double(int)> &y);
+  inline std::pair<double, double> minmax(int left, int right) const { return get_minmax(1, 0, size - 1, left, right); }
 
+private:
+  std::pair<double, double> get_minmax(int n, int left, int right, int range_left, int range_right) const;
+  void build_tree(const std::function<double(int)> &y, int n, int left, int right);
+  std::vector<std::pair<double, double>> tree;
+  int size = 0;
+};
+
+// maps a linear slider position onto a log10 scale
+class LogScale {
+public:
+  LogScale(double factor) : factor(factor) {}
   void setRange(double min, double max) {
     log_min = factor * std::log10(min);
     log_max = factor * std::log10(max);
-    QSlider::setRange(min, max);
-    setValue(QSlider::value());
   }
-  int value() const {
-    double v = log_min + (log_max - log_min) * ((QSlider::value() - minimum()) / double(maximum() - minimum()));
+  int value(int pos, int pos_min, int pos_max) const {
+    double v = log_min + (log_max - log_min) * ((pos - pos_min) / double(pos_max - pos_min));
     return std::lround(std::pow(10, v / factor));
   }
-  void setValue(int v) {
+  int position(int v, int pos_min, int pos_max) const {
     double log_v = std::clamp(factor * std::log10(v), log_min, log_max);
-    v = minimum() + (maximum() - minimum()) * ((log_v - log_min) / (log_max - log_min));
-    QSlider::setValue(v);
+    return pos_min + (pos_max - pos_min) * ((log_v - log_min) / (log_max - log_min));
   }
 
 private:
   double factor, log_min = 0, log_max = 1;
 };
 
-enum {
-  ColorsRole = Qt::UserRole + 1,
-  BytesRole = Qt::UserRole + 2
-};
+enum class ValidState { Invalid, Intermediate, Acceptable };
 
-class SegmentTree {
-public:
-  SegmentTree() = default;
-  void build(const std::vector<QPointF> &arr);
-  inline std::pair<double, double> minmax(int left, int right) const { return get_minmax(1, 0, size - 1, left, right); }
+// single identifier: one or more [A-Za-z0-9_], spaces rewritten to '_'
+ValidState validateName(std::string &input);
+// comma-separated identifiers: \w+(,\w+)*
+ValidState validateNodes(const std::string &input);
+// one or more non-whitespace characters (\S+)
+ValidState validateNonWhitespace(const std::string &input);
+// dotted IPv4 address (0-255 per octet)
+ValidState validateIpAddress(const std::string &input);
+// C-locale floating-point
+ValidState validateDouble(const std::string &input);
 
-private:
-  std::pair<double, double> get_minmax(int n, int left, int right, int range_left, int range_right) const;
-  void build_tree(const std::vector<QPointF> &arr, int n, int left, int right);
-  std::vector<std::pair<double, double>> tree;
-  int size = 0;
-};
-
-class MessageBytesDelegate : public QStyledItemDelegate {
-  Q_OBJECT
-public:
-  MessageBytesDelegate(QObject *parent, bool multiple_lines = false);
-  void paint(QPainter *painter, const QStyleOptionViewItem &option, const QModelIndex &index) const override;
-  QSize sizeHint(const QStyleOptionViewItem &option, const QModelIndex &index) const override;
-  bool multipleLines() const { return multiple_lines; }
-  void setMultipleLines(bool v) { multiple_lines = v; }
-  QSize sizeForBytes(int n) const;
-
-private:
-  std::array<QStaticText, 256> hex_text_table;
-  QFontMetrics font_metrics;
-  QFont fixed_font;
-  QSize byte_size = {};
-  bool multiple_lines = false;
-  int h_margin, v_margin;
-};
-
-// Accepts a single identifier: one or more [A-Za-z0-9_], spaces rewritten to '_'.
-class NameValidator : public QValidator {
-  Q_OBJECT
-public:
-  NameValidator(QObject *parent=nullptr);
-  QValidator::State validate(QString &input, int &pos) const override;
-};
-
-// Accepts comma-separated identifiers: \w+(,\w+)*
-class NodeValidator : public QValidator {
-  Q_OBJECT
-public:
-  NodeValidator(QObject *parent=nullptr);
-  QValidator::State validate(QString &input, int &pos) const override;
-};
-
-// Accepts one or more non-whitespace characters (\S+).
-class NonWhitespaceValidator : public QValidator {
-  Q_OBJECT
-public:
-  NonWhitespaceValidator(QObject *parent=nullptr);
-  QValidator::State validate(QString &input, int &pos) const override;
-};
-
-// Accepts a dotted IPv4 address (0-255 per octet).
-class IpAddressValidator : public QValidator {
-  Q_OBJECT
-public:
-  IpAddressValidator(QObject *parent=nullptr);
-  QValidator::State validate(QString &input, int &pos) const override;
-};
-
-// C-locale floating-point validator (matches QString::toDouble).
-class DoubleValidator : public QValidator {
-  Q_OBJECT
-public:
-  DoubleValidator(QObject *parent = nullptr);
-  QValidator::State validate(QString &input, int &pos) const override;
+// "Darcula" like dark theme
+struct DarkTheme {
+  static constexpr CabanaColor window{0x35, 0x35, 0x35};
+  static constexpr CabanaColor window_text{0xbb, 0xbb, 0xbb};
+  static constexpr CabanaColor base{0x3c, 0x3f, 0x41};
+  static constexpr CabanaColor tooltip_text{0xbb, 0xbb, 0xbb};
+  static constexpr CabanaColor text{0xbb, 0xbb, 0xbb};
+  static constexpr CabanaColor button{0x3c, 0x3f, 0x41};
+  static constexpr CabanaColor highlight{0x2f, 0x65, 0xca};
+  static constexpr CabanaColor bright_text{0xf0, 0xf0, 0xf0};
+  static constexpr CabanaColor disabled_text{0x77, 0x77, 0x77};
+  static constexpr CabanaColor light{0x77, 0x77, 0x77};
+  static constexpr CabanaColor dark{0x35, 0x35, 0x35};
 };
 
 namespace utils {
@@ -148,67 +80,23 @@ bool isMainThread();
 // inline on the main thread, queued until drainMainThreadQueue() otherwise
 void runOnMainThread(std::function<void()> fn);
 void drainMainThreadQueue();
-QPixmap icon(const QString &id);
 std::string homePath();
 std::filesystem::path configPath();
 bool getClipboardText(std::string *text);  // false if no clipboard tool is available
 bool setClipboardText(const std::string &text);
-bool isDarkTheme();
-void setTheme(int theme);
-inline void drawStaticText(QPainter *p, const QRect &r, const QStaticText &text) {
-  auto size = (r.size() - text.size()) / 2;
-  p->drawStaticText(r.left() + size.width(), r.top() + size.height(), text);
-}
+std::string bootstrapSvg(const std::string &id);  // empty if unknown
 
 // boundary conversions for the remaining Qt byte-array based state APIs
 template <typename T>
 std::vector<uint8_t> toBytes(const T &dat) { return {dat.begin(), dat.end()}; }
-inline auto qbytes(const std::vector<uint8_t> &dat) {
-  return decltype(QString().toUtf8())((const char *)dat.data(), (int)dat.size());
-}
 
 }
 
-class ToolButton : public QToolButton {
-  Q_OBJECT
-public:
-  ToolButton(const QString &icon, const QString &tooltip = {}, QWidget *parent = nullptr) : QToolButton(parent) {
-    setIcon(icon);
-    setToolTip(tooltip);
-    setAutoRaise(true);
-    const int metric = QApplication::style()->pixelMetric(QStyle::PM_SmallIconSize);
-    setIconSize({metric, metric});
-    theme = settings.theme;
-    settings_connection_ = settings.changed.connect([this]() { updateIcon(); });
-  }
-  void setIcon(const QString &icon) {
-    icon_str = icon;
-    QToolButton::setIcon(utils::icon(icon_str));
-  }
-
-private:
-  void updateIcon() { if (std::exchange(theme, settings.theme) != theme) setIcon(icon_str); }
-  Connection settings_connection_;
-  QString icon_str;
-  int theme;
-};
-
-class TabBar : public QTabBar {
-  Q_OBJECT
-
-public:
-  TabBar(QWidget *parent) : QTabBar(parent) {}
-  int addTab(const QString &text);
-
-private:
-  void closeTabClicked();
-};
-
-// Watches SIGINT/SIGTERM via a self-pipe and a dedicated waiter thread
-// (no Qt notifiers/timers). Exit is marshaled onto the GUI thread.
+// Watches SIGINT/SIGTERM via a self-pipe and a dedicated waiter thread.
+// on_signal runs on the waiter thread; the caller marshals to the GUI thread.
 class UnixSignalHandler {
 public:
-  UnixSignalHandler();
+  UnixSignalHandler(std::function<void()> on_signal);
   ~UnixSignalHandler();
   static void signalHandler(int s);
 
@@ -220,5 +108,3 @@ private:
 
 int num_decimals(double num);
 std::filesystem::path executableDir();
-void initApp(int argc, char *argv[], bool disable_hidpi = true);
-QPixmap bootstrapPixmap(const QString &id);
